requirements.txt
Django==1.9.5
django-filter==0.13.0
django-tables2==1.1.6
django-tables2-reports==0.0.9
djangorestframework==3.3.3
MySQL-python==1.2.5
six==1.10.0
South==1.0
wheel==0.24.0
django_tables2